move virtual example plugin to a standalone package
