+++ packages/editor/package.json
@@ -38,6 +38,7 @@
   },
   "devDependencies": {
     "@babel/preset-react": "^7.14.5",
+    "@meta-ui/vite-plugins": "^1.0.0",
     "@types/codemirror": "^5.60.5",
     "@types/lodash-es": "^4.17.5",
     "@vitejs/plugin-react": "^1.0.1",
