+++ packages/editor/playground.html
@@ -1,0 +1,15 @@
+<!DOCTYPE html>
+<html lang="en" style="overflow: hidden">
+  <head>
+    <meta charset="UTF-8" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>meta-ui playground</title>
+  </head>
+  <body>
+    <div id="root"></div>
+    <script type="module">
+      import renderApp from './src/main.tsx';
+      renderApp();
+    </script>
+  </body>
+</html>
