+++ packages/editor/vite.config.ts
@@ -1,5 +1,6 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
+import { virtualExamplePlugin } from '@meta-ui/vite-plugins';
 
 // https://vitejs.dev/config/
 export default defineConfig({
@@ -20,6 +21,7 @@ export default defineConfig({
         ],
       },
     }),
+    virtualExamplePlugin(),
   ],
   define: {
     // https://github.com/satya164/react-simple-code-editor/issues/86
