+++ packages/runtime/package.json
@@ -58,6 +58,7 @@
     "@babel/preset-env": "^7.15.6",
     "@babel/preset-react": "^7.14.5",
     "@babel/preset-typescript": "^7.15.0",
+    "@meta-ui/vite-plugins": "^1.0.0",
     "@testing-library/react": "^12.1.0",
     "@types/lodash": "^4.14.170",
     "@types/lodash-es": "^4.17.5",
