+++ packages/vite-plugins/README.md
@@ -1,0 +1,11 @@
+# `@meta-ui/vite-plugins`
+
+> TODO: description
+
+## Usage
+
+```
+const vitePlugins = require('@meta-ui/vite-plugins');
+
+// TODO: DEMONSTRATE API
+```
